+++ prestashop-v8/scalexpertplugin/controllers/front/ajax.php
@@ -8,71 +8,35 @@
  * @license   https://opensource.org/licenses/osl-3.0.php Open Software License (OSL 3.0)
  */
 
-use ScalexpertPlugin\Form\Configuration\FinancingConfigurationFormDataConfiguration;
-use ScalexpertPlugin\Form\Customize\DesignCustomizeFormDataConfiguration;
+use ScalexpertPlugin\Service\AvailableSolutionsService;
 
 class ScalexpertpluginAjaxModuleFrontController extends ModuleFrontController
 {
-    public function displayAjaxGetFinancialInsertsOnProduct()
+    public function displayAjaxGetFinancingSimulationInsertOnProduct()
     {
         $productID = Tools::getValue('id_product', null);
-        $financialInserts = [];
+        $productAttributeID = Tools::getValue('id_product_attribute', null);
 
         if (!empty($productID)) {
+            /* @var AvailableSolutionsService $availableSolutionsService */
             $availableSolutionsService = $this->get('scalexpert.service.available_solutions');
-            $availableFinancialSolutions = $availableSolutionsService->getAvailableFinancialSolutions($productID);
+            $buyerBillingCountry = $availableSolutionsService->getContextBuyerBillingCountry();
 
-            if (!empty($availableFinancialSolutions)) {
-                $sortedSolutions = [];
-                foreach ($availableFinancialSolutions as $availableFinancialSolution) {
-                    $financialSolutionData = $availableFinancialSolution;
-
-                    if (!empty($financialSolutionData['designConfiguration'])) {
-                        if (isset($financialSolutionData['designConfiguration']['productDisplay']) && !$financialSolutionData['designConfiguration']['productDisplay']) {
-                            continue;
-                        }
-
-                        if (isset($financialSolutionData['designConfiguration']['productDisplayLogo'])) {
-                            $financialSolutionData['displayLogo'] = $financialSolutionData['designConfiguration']['productDisplayLogo'];
-                        }
-
-                        if (isset($financialSolutionData['designConfiguration']['productPosition'])) {
-                            $financialSolutionData['position'] = $financialSolutionData['designConfiguration']['productPosition'];
-                        }
-
-                        if (isset($financialSolutionData['designConfiguration']['productTitle'])) {
-                            $financialSolutionData['visualTitle'] = $financialSolutionData['designConfiguration']['productTitle'];
-                        }
-
-                        if (isset($financialSolutionData['designConfiguration']['position'])) {
-                            $financialSolutionData['position'] = $financialSolutionData['designConfiguration']['position'];
-                        }
-                    }
-
-                    $sortedSolutions[] = $financialSolutionData;
-                }
-
-                $this->module->sortSolutionsByPosition($sortedSolutions);
-
-                foreach ($sortedSolutions as $sortedSolution) {
-                    $this->context->smarty->assign(['financialSolution' => $sortedSolution]);
-                    $formattedInsert = $this->context->smarty->fetch(
-                        'module:' . $this->module->name . '/views/templates/hook/financial-insert.tpl'
-                    );
-
-                    if (empty($sortedSolution['productPosition'])) {
-                        $sortedSolution['productPosition'] = 'displayProductAdditionalInfo';
-                    }
-
-                    $financialInserts[] = [
-                        'formattedInsert' => $formattedInsert,
-                        'hook' => $sortedSolution['productPosition'],
-                    ];
-                }
+            if ('fr' === $buyerBillingCountry) {
+                $simulationInsert = $this->_getSimulationInsert(
+                    $availableSolutionsService,
+                    $productID,
+                    $productAttributeID
+                );
+                $this->ajaxDie(json_encode(['simulationInsert' => $simulationInsert]));
+            } else {
+                $financialInserts = $this->_getFinancialInserts(
+                    $availableSolutionsService,
+                    $productID
+                );
+                $this->ajaxDie(json_encode(['financialInserts' => $financialInserts]));
             }
         }
-
-        $this->ajaxDie(json_encode(['financialInserts' => $financialInserts]));
     }
 
     public function displayAjaxGetInsuranceInserts()
@@ -96,8 +60,12 @@ class ScalexpertpluginAjaxModuleFrontController extends ModuleFrontController
             $pageContext = 'cart';
         }
 
+        /* @var AvailableSolutionsService $availableSolutionsService */
         $availableSolutionsService = $this->get('scalexpert.service.available_solutions');
-        $availableInsuranceSolutions = $availableSolutionsService->getAvailableInsuranceSolutions($productID, $productIDAttribute);
+        $availableInsuranceSolutions = $availableSolutionsService->getAvailableInsuranceSolutions(
+            $productID,
+            $productIDAttribute
+        );
 
         if (!empty($availableInsuranceSolutions)) {
             foreach ($availableInsuranceSolutions as $availableInsuranceSolution) {
@@ -106,20 +74,26 @@ class ScalexpertpluginAjaxModuleFrontController extends ModuleFrontController
                 $insuranceSolutionData['id_product_attribute'] = $productIDAttribute;
 
                 if (!empty($insuranceSolutionData['designConfiguration'])) {
-                    if (isset($insuranceSolutionData['designConfiguration'][$pageContext . '_display']) && !$insuranceSolutionData['designConfiguration'][$pageContext . '_display']) {
+                    if (
+                        isset($insuranceSolutionData['designConfiguration'][$pageContext . '_display'])
+                        && !$insuranceSolutionData['designConfiguration'][$pageContext . '_display']
+                    ) {
                         continue;
                     }
 
                     if (isset($insuranceSolutionData['designConfiguration'][$pageContext . '_display_logo'])) {
-                        $insuranceSolutionData['displayLogo'] = $insuranceSolutionData['designConfiguration'][$pageContext . '_display_logo'];
+                        $insuranceSolutionData['displayLogo'] =
+                            $insuranceSolutionData['designConfiguration'][$pageContext . '_display_logo'];
                     }
 
                     if (isset($insuranceSolutionData['designConfiguration'][$pageContext . '_position'])) {
-                        $insuranceSolutionData['position'] = $insuranceSolutionData['designConfiguration'][$pageContext . '_position'];
+                        $insuranceSolutionData['position'] =
+                            $insuranceSolutionData['designConfiguration'][$pageContext . '_position'];
                     }
 
                     if (isset($insuranceSolutionData['designConfiguration'][$pageContext . '_title'])) {
-                        $insuranceSolutionData['visualTitle'] = $insuranceSolutionData['designConfiguration'][$pageContext . '_title'];
+                        $insuranceSolutionData['visualTitle'] =
+                            $insuranceSolutionData['designConfiguration'][$pageContext . '_title'];
                     }
                 }
 
@@ -145,5 +119,128 @@ class ScalexpertpluginAjaxModuleFrontController extends ModuleFrontController
         }
 
         $this->ajaxRender(json_encode(['insuranceInserts' => $insuranceInserts]));
+    }
+
+    private function _getSimulationInsert(
+        AvailableSolutionsService $availableSolutionsService,
+        $productID,
+        $productAttributeID
+    ): string
+    {
+        $simulationInsert = '';
+        $availableSimulation = $availableSolutionsService->getSimulationForAvailableFinancialSolutions(
+            $productID,
+            $productAttributeID
+        );
+
+        if (
+            !empty($availableSimulation)
+            && isset($availableSimulation['solutionSimulations'])
+        ) {
+            // Group financing solutions by having fees or not
+            $groupedSolutionSimulations = [];
+            foreach ($availableSimulation['solutionSimulations'] as $solutionSimulation) {
+                foreach ($solutionSimulation['simulations'] as $simulation) {
+                    $simulation['designConfiguration'] = $solutionSimulation['designConfiguration'];
+                    $simulation['isLongFinancingSolution'] = $solutionSimulation['isLongFinancingSolution'];
+                    $simulation['hasFeesOnFirstInstallment'] =
+                        $solutionSimulation['hasFeesSolution']
+                        && 0 < $simulation['feesAmount']
+                    ;
+
+                    $groupedSolutionSimulations['all'][] = $simulation;
+                }
+            }
+
+            // Sort by duration
+            if (!empty($groupedSolutionSimulations)) {
+                $this->sortSolutionsByDuration($groupedSolutionSimulations['all']);
+            }
+
+            $this->context->smarty->assign([
+                'solutionSimulations' => $groupedSolutionSimulations,
+                'financedAmount' => $availableSimulation['financedAmount'] ?? '',
+                'financedAmountFormatted' => $availableSimulation['financedAmountFormatted'] ?? '',
+            ]);
+
+            $simulationInsert = $this->context->smarty->fetch(
+                'module:' . $this->module->name . '/views/templates/hook/simulation-insert.tpl'
+            );
+        }
+
+        return $simulationInsert;
+    }
+
+    private function _getFinancialInserts(
+        AvailableSolutionsService $availableSolutionsService,
+        $productID
+    ): array
+    {
+        $financialInserts = [];
+        $availableFinancialSolutions = $availableSolutionsService->getAvailableFinancialSolutions($productID);
+
+        if (!empty($availableFinancialSolutions)) {
+            $sortedSolutions = [];
+            foreach ($availableFinancialSolutions as $availableFinancialSolution) {
+                $financialSolutionData = $availableFinancialSolution;
+
+                if (!empty($financialSolutionData['designConfiguration'])) {
+                    if (
+                        isset($financialSolutionData['designConfiguration']['productDisplay'])
+                        && !$financialSolutionData['designConfiguration']['productDisplay']
+                    ) {
+                        continue;
+                    }
+
+                    if (isset($financialSolutionData['designConfiguration']['productDisplayLogo'])) {
+                        $financialSolutionData['displayLogo'] =
+                            $financialSolutionData['designConfiguration']['productDisplayLogo'];
+                    }
+
+                    if (isset($financialSolutionData['designConfiguration']['productPosition'])) {
+                        $financialSolutionData['position'] =
+                            $financialSolutionData['designConfiguration']['productPosition'];
+                    }
+
+                    if (isset($financialSolutionData['designConfiguration']['productTitle'])) {
+                        $financialSolutionData['visualTitle'] =
+                            $financialSolutionData['designConfiguration']['productTitle'];
+                    }
+
+                    if (isset($financialSolutionData['designConfiguration']['position'])) {
+                        $financialSolutionData['position'] = $financialSolutionData['designConfiguration']['position'];
+                    }
+                }
+
+                $sortedSolutions[] = $financialSolutionData;
+            }
+
+            $this->module->sortSolutionsByPosition($sortedSolutions);
+
+            foreach ($sortedSolutions as $sortedSolution) {
+                $this->context->smarty->assign(['financialSolution' => $sortedSolution]);
+                $formattedInsert = $this->context->smarty->fetch(
+                    'module:' . $this->module->name . '/views/templates/hook/financial-insert.tpl'
+                );
+
+                if (empty($sortedSolution['productPosition'])) {
+                    $sortedSolution['productPosition'] = 'displayProductAdditionalInfo';
+                }
+
+                $financialInserts[] = [
+                    'formattedInsert' => $formattedInsert,
+                    'hook' => $sortedSolution['productPosition'],
+                ];
+            }
+        }
+
+        return $financialInserts;
+    }
+
+    private function sortSolutionsByDuration(&$solutions): void
+    {
+        uasort($solutions, function ($a, $b) {
+            return $a['duration'] > $b['duration'];
+        });
     }
 }